Rewrite menu item decoration
The underline now belongs to the clickable item, so we can use :hover to
style it. Markup needs to be changed a little to incorporate the styling
for :checked inputs, but otherwise only (a lot of) positioning tweaks are
involved.

The menu-item mixin is rewritten to work better. The volatile
> *:first-child selector is removed to favour > a, > label (mainly so we
can move elements around to make ~ work), and therefore certain selectors
are extracted into variable, and the menu-item mixin is converted into a
block so it can be better reused.