Implement hover-to-open menus
So the submenu wrapper works! Hooray.

The hover effect is implemented by auto-checking the checkbox when
hovering. The unchecking requires more work because we need to prevent
it when going from the parent to a child (i.e. from top menu item to the
submenu). But it works, and is reletively straightforward in the end.

Also add an opacity transform so the effect looks better when hovering
between multiple menu items (the submenu won't look overlapping).

Still strictly progressive. I'm happy.